Added Quests, which behave like goals did previously, except that they are composed of a List of goals that must be completed in linear order. Quests have prereq and postreq Quests The Plot class now saves active and completed quests via preferences

+++ src/com/arnopaja/supermac/world/grid/Location.java
@@ -23,7 +23,7 @@ public class Location {
         this.facing = facing;
     }
 
-    public RenderGrid getRenderGrid(float renderGridWidth, float renderGridHeight) {
+    public RenderGrid getRenderGrid(int renderGridWidth, int renderGridHeight) {
         return grid.getRenderGrid(position, renderGridWidth, renderGridHeight);
     }
